+++ src/app/builder/(components)/DetailedBuildView.tsx
@@ -39,9 +39,9 @@ function MasonryCard({ data: item, onMoreInfoClick }: MasonryBuildItem) {
             className="mx-auto mb-2 h-32 w-32 flex-shrink-0 rounded-full"
             width={200}
             height={200}
-            src={`https://${process.env.NEXT_PUBLIC_IMAGE_URL}${imagePath}?width=200&height=200`}
+            src={`https://${process.env.NEXT_PUBLIC_IMAGE_URL}${imagePath}`}
             alt={`${name} icon`}
-            priority={true}
+            loading="eager"
           />
 
           {name}