+++ src/app/page.tsx
@@ -13,6 +13,7 @@ export default function HomePage() {
           alt="Home page background"
           layout="fill"
           className="absolute inset-0 -z-10 h-full w-full object-cover object-right md:object-center"
+          loading="eager"
         />
         <div className="sm:transform-g1 hidden sm:absolute sm:-top-10 sm:right-1/2 sm:-z-10 sm:mr-10 sm:block sm:blur-3xl">
           <div