+++ src/app/tracker/(components)/ItemCard.tsx
@@ -17,11 +17,11 @@ function CardImage({ item }: { item: ItemCardProps['item'] }) {
     <div className="relative flex h-[64px] w-full grow items-center justify-center overflow-hidden bg-[url('https://d2sqltdcj8czo5.cloudfront.net/card-body-bg.jpg')]">
       {item && (
         <Image
-          src={`https://${process.env.NEXT_PUBLIC_IMAGE_URL}${item.imagePath}?width=${imageSize.width}&height=${imageSize.height}`}
+          src={`https://${process.env.NEXT_PUBLIC_IMAGE_URL}${item.imagePath}`}
           alt={item.name}
           width={imageSize.width}
           height={imageSize.height}
-          priority={true}
+          loading="eager"
         />
       )}
     </div>